server/requirements.txt
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.
django-prometheus==2.3.1
django-filter==24.3
psycopg==3.2.3
psycopg2-binary==2.9.10
stripe==11.2.0
python-decouple==3.8
drf-yasg==1.21.8
reportlab==4.2.5